Plug link detection in add_link method
Before adding a link, check that it would not cross any other ones.
If that is the case, do not add it and return False.

Note: this may not be time efficient and it may be better to add a
full check only at the end (when we check if the solution is ok).